fixed IE compatibility problems
I am unsure if this would effect other users, but my IE browser was forcing content to load in IE8 document mode, so I added a meta statement setting the IE X-UA-Compatible http-equiv to 'edge' - this solved the problem on my end. The page now renders all HTML5 elements as intended in IE and flows responsively.